Found a regression in /play. Need to fix.
Fix chatgpt module and break some things
