+++ crack-config/src/lib.rs
@@ -94,9 +94,9 @@ pub async fn poise_framework(
             commands::autopause(),
             // commands::boop(),
             commands::coinflip(),
-            // //commands::create_playlist(),
-            // //commands::delete_playlist(),
-            // //commands::chatgpt(),
+            //commands::create_playlist(),
+            //commands::delete_playlist(),
+            commands::chatgpt(),
             commands::clear(),
             commands::help(),
             commands::leave(),
