+++ crack-core/src/commands/admin/set_all_log_channel.rs
@@ -1,34 +1,27 @@
-use serenity::all::Channel;
-
 use crate::{
-    commands::set_all_log_channel_old_data, messaging::message::CrackedMessage,
+    commands::set_all_log_channel_data, messaging::message::CrackedMessage,
     utils::send_response_poise, Context, Error,
 };
+use serenity::all::Channel;
+
 /// Set the join-leave log channel.
 #[poise::command(prefix_command, owners_only)]
 pub async fn set_all_log_channel(
     ctx: Context<'_>,
-    #[description = "Channel to send all logs"] channel: Channel,
+    #[description = "Channel to send all logs"] channel: Option<Channel>,
+    #[description = "ChannelId to send all logs"] channel_id: Option<
+        serenity::model::id::ChannelId,
+    >,
 ) -> Result<(), Error> {
-    let channel_id = channel.id();
+    let channel_id = if let Some(channel) = channel {
+        channel.id()
+    } else {
+        channel_id.unwrap()
+    };
     let guild_id = ctx.guild_id().unwrap();
-    // let mut data = ctx.serenity_context().data.write().await;
-    // let data = &ctx.serenity_context().data;
 
-    set_all_log_channel_old_data(ctx.serenity_context().data.clone(), guild_id, channel_id).await?;
-    // set_all_log_channel_old_data(ctx.data().clone(), guild_id, channel_id).await?;
-    // let _entry = &data
-    //     .get_mut::<GuildSettingsMap>()
-    //     .unwrap()
-    //     .entry(guild_id)
-    //     .and_modify(|e| e.set_all_log_channel(channel_id.get()));
-
-    // let settings = data
-    //     .get_mut::<GuildSettingsMap>()
-    //     .unwrap()
-    //     .get_mut(&guild_id);
-
-    // let _res = settings.map(|s| s.save()).unwrap();
+    // set_all_log_channel_old_data(ctx.serenity_context().data.clone(), guild_id, channel_id).await?;
+    set_all_log_channel_data(ctx.data(), guild_id, channel_id).await?;
 
     send_response_poise(
         ctx,
