+++ crack-core/src/commands/chatgpt.rs
