+++ crack-core/src/commands/mod.rs
@@ -1,12 +1,12 @@
 pub mod admin;
-// pub mod crack_gpt;
+pub mod chatgpt;
 pub mod music;
 pub mod ping;
 // pub mod playlist;
 pub mod version;
 
 pub use admin::*;
-// pub use crack_gpt::*;
+pub use chatgpt::*;
 pub use music::*;
 pub use ping::*;
 // pub use playlist::*;
